demographics work now it can add data about which races and languages exist in the area, and it will slowly deviate the data as it goes down into child objects. There's a chance that a given stat might entirely be replaced for something else. For instance, for races, you might have a collection of 5 species of different percentages. By the next iteration, one of the species will get swapped out from a wookie at 18% into a hutt at 49%.
The percentages of 'swaps' do not reference each other, it just picks a fully random number. The rest
of the stats DO reference their own previous values, only deviating slightly from them.

Currently, there is a bug in which on the subsequent iterations, the percentages don't add up to 100%,
and some values can even become negative. This is because the diff value isn't being accounted for
in the newStats thing (when you replace a Wookie with a Hutt).